+++ src/main/java/com/example/astonrest/service/UserService.java
@@ -2,8 +2,10 @@ package com.example.astonrest.service;
 
 import com.example.astonrest.dto.UserDTO;
 import com.example.astonrest.entity.User;
+import com.example.astonrest.exception.NotFoundException;
 import com.example.astonrest.mapper.UserMapper;
 import com.example.astonrest.repository.UserRepository;
+import com.example.astonrest.util.UserValidator;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -20,6 +22,7 @@ public class UserService {
      * @param userDTO DTO пользователя
      */
     public void createUser(UserDTO userDTO) {
+        UserValidator.validate(userDTO);
         User user = UserMapper.toEntity(userDTO);
         userRepository.save(user);
     }
@@ -67,7 +70,10 @@ public class UserService {
      * @param id ID пользователя
      */
     public void deleteUser(int id) {
+        UserDTO user = getUserById(id);
+        if(user == null) {
+            throw new NotFoundException("User with ID " + id + " not found.");
+        }
         userRepository.delete(id);
-
     }
 }
